removing Base classes fixed project file -> RTFTextParser used new implementation but old package
