+++ ToxyFramework/ToxyFramework.csproj
@@ -46,7 +46,7 @@
     <PackageReference Include="NPOI" Version="2.7.5" />
     <PackageReference Include="PasswordProtectedChecker" Version="2.1.3" />
     <PackageReference Include="PdfPig" Version="0.1.13" />
-    <PackageReference Include="ReasonableRTF.Standard" Version="2025.11.1" />
+    <PackageReference Include="ReasonableRTF.Standard" Version="2026.4.0" />
     <PackageReference Include="Tabula" Version="0.1.5" />
     <PackageReference Include="taglib-sharp-netstandard2.0" Version="2.1.0" />
     <PackageReference Include="VCardReader" Version="1.2.3" />
